[wip] Introduce macros for registering associations
