+++ tests/sio_io/write_frame_sio.cpp
@@ -1,11 +1,6 @@
-#include "datamodel/ExampleWithARelation.h"
 #include "write_frame.h"
 
 #include "podio/SIOWriter.h"
-
-#include "podio/AssociationSIOBlock.h"
-
-const static auto foo = podio::AssociationSIOBlock<ExampleMC, ex42::ExampleWithARelation>{};
 
 int main(int, char**) {
   write_frames<podio::SIOWriter>("example_frame.sio");
